# frontend/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Frontend",
        "version": "v0.0.0"
    },
    "paths": {
...
        "/api/frontend/v1/clubs/clubs": {
            "get": {
                "tags": [
                    "Clubs"
                ],
                "summary": "Get all clubs",
                "description": "Get all clubs",
                "operationId": "get_all_clubs",
                "responses": {
                    "400": {
                        "description": "Client side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClubList"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ApiErrorResponse": {
                "description": "The response that is sent in a case of an error",
                "type": "object",
                "properties": {
                    "message": {
                        "description": "A human-readable error message.\n\nMay be used for displaying purposes",
                        "type": "string"
                    },
                    "status_code": {
                        "$ref": "#/components/schemas/ApiStatusCode"
                    }
                },
                "required": [
                    "message",
                    "status_code"
                ]
            },
            "ApiStatusCode": {
                "description": "The Status code that are returned throughout the API",
                "type": "integer",
                "enum": [
                    1000,
                    1001,
                    1002,
                    1003,
                    2000
                ]
            },
...
            "ClubList": {
                "description": "A list of clubs",
                "type": "object",
                "properties": {
                    "clubs": {
                        "description": "List of all clubs",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FullClub"
                        }
                    }
                },
                "required": [
                    "clubs"
                ]
            },
...
            "FullClub": {
                "description": "A full representation of a club",
                "type": "object",
                "properties": {
                    "name": {
                        "description": "Name of the club",
                        "type": "string"
                    },
                    "users": {
                        "description": "User count associated with the club",
                        "type": "integer",
                        "format": "uint64",
                        "minimum": 0
                    },
                    "uuid": {
                        "description": "Primary key",
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "name",
                    "users",
                    "uuid"
                ]
            },
...
        }
    }
}